Refacor(#97): cycle 검증을 도메인 계층에서 하도록 변경
